# refactor(parser): parse quoted text attribute once