Do not repeat parsing of attributes for each type of quoted text

Fixes #827

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>