@xcoulon 2021-02-28
FWIW, the `benchstat` results when running `make bench` on this branch compared to `master`:
```
name           old time/op    new time/op    delta
Libasciidoc-8     135ms ± 1%     129ms ± 6%   -4.58%  (p=0.001 n=10+9)

name           old alloc/op   new alloc/op   delta
Libasciidoc-8    52.4MB ± 0%    44.6MB ± 0%  -14.96%  (p=0.000 n=10+10)

name           old allocs/op  new allocs/op  delta
Libasciidoc-8      903k ± 0%      837k ± 0%   -7.32%  (p=0.000 n=9+10)
```